Moved all constants to a separate file
