+++ gameBoard.py
@@ -2,16 +2,7 @@ from graphics import Window, Point, WINDOW_HEIGHT, WINDOW_WIDTH, BG_COL, LINE_WI
 from tkinter import Tk
 from PIL import ImageTk, Image
 from userInterface import UserInterface, SPRITE_BUFFER, do_nothing
-
-DEFAULT_SQUARE_SIZE = 64 + SPRITE_BUFFER
-SELECTION_BUFFER = 3
-SELECTION_SQUARE = DEFAULT_SQUARE_SIZE - SELECTION_BUFFER
-BOARD_ROWS = 8
-BOARD_COLS = 8
-BOARD_WIDTH = DEFAULT_SQUARE_SIZE * BOARD_COLS
-BOARD_HEIGHT = DEFAULT_SQUARE_SIZE * BOARD_ROWS
-DEFAULT_X_POS = (WINDOW_WIDTH - BOARD_WIDTH) // 2
-DEFAULT_Y_POS = (WINDOW_HEIGHT - BOARD_HEIGHT) // 4
+from constants import *
 
 
 class GameBoard:
