+++ units.py
@@ -4,38 +4,7 @@ from gameBoard import Space, BOARD_ROWS, BOARD_COLS
 from graphics import SpriteType
 from random import randint
 from names import Names, Titles
-
-FIRST_STRIKE_BOOST = 1.2
-POOR_EFFECT_MOD = 4/5
-STRONG_EFFECT_MOD = 5/4
-
-class DamageType(IntEnum):
-    SLASH = 1
-    PIERCE = 2
-    BLUDGEON = 3
-    MAGIC = 4
-
-class ArmourType(IntEnum):
-    ROBES = 1
-    PADDED = 2
-    CHAIN = 3
-    PLATE = 4
-
-class Effect(IntEnum):
-    POOR = 1
-    NEUTRAL = 2
-    STRONG = 3
-
-class MoveSpeed(IntEnum):
-    SLOW = 2
-    MED = 3
-    FAST = 4
-
-class MoveType(IntEnum):
-    FOOT = 1
-    HORSE = 2
-    FLY = 3
-
+from constants import *
 
 class Unit:
     def __init__(
@@ -335,6 +304,72 @@ class Soldier(Unit):
         #target_spaces = target_spaces.union(self.check_target_spaces(i-1, j+1, range, space_list))
         return target_spaces
 
+class Archer(Unit):
+    def __init__(self, p1 = True) -> None:
+        hp=15
+        dam_val=6
+        dam_type=DamageType.PIERCE
+        arm_val=2
+        arm_type=ArmourType.PADDED
+        move=MoveSpeed.MED
+        move_type = MoveType.FOOT
+        if p1:
+            sprite = SpriteType.ARCHER1
+        else:
+            sprite = SpriteType.ARCHER2
+        name_list = Names.Commoner
+        title_list = Titles.Archer
+        ability_name = "Ranged Attack"
+        ability_range = 5
+        ability_value = 7
+        super().__init__(hp, dam_val, dam_type, arm_val, arm_type, move, move_type, sprite, name_list, title_list, ability_name, ability_range, ability_value)
+        self.__special_damage_type = DamageType.PIERCE
+
+    def special_ability(self, target, spaces):
+        unit_name = self.get_name()
+        target_name = target.get_name()
+        attack_log = []
+        first_strike_attack = ceil(self.get_ability_value() * FIRST_STRIKE_BOOST)
+        target_hp = target.get_curr_hp()
+        self.attack(target, first_strike_attack, self.__special_damage_type)
+        damage_dealt = target_hp - target.get_curr_hp()
+        attack_log.append(f"{unit_name} fires an arrow at {target_name}, dealing {damage_dealt} damage!\n")
+        if target.is_dead():
+            attack_log.append(f"{unit_name} has slain {target_name}!\n")
+            target.get_location().assign_unit(None)
+        return attack_log
+
+class Cavalry(Unit):
+    def __init__(self, p1 = True) -> None:
+        hp=20
+        dam_val=9
+        dam_type=DamageType.SLASH
+        arm_val=4
+        arm_type=ArmourType.PLATE
+        move=MoveSpeed.FAST
+        move_type = MoveType.HORSE
+        if p1:
+            sprite = SpriteType.PEASANT1
+        else:
+            sprite = SpriteType.PEASANT2
+        name_list = Names.Noble
+        title_list = Titles.Cavalry
+        ability_name = "Harrying Strike"
+        ability_range = 1
+        ability_value = None
+        super().__init__(hp, dam_val, dam_type, arm_val, arm_type, move, move_type, sprite, name_list, title_list, ability_name, ability_range, ability_value)
+    
+    def check_move_spaces(self, i: int, j: int, range: int, space_list: list) -> set:
+        valid_spaces = set()
+        if i >= 0 and i < BOARD_ROWS and j >= 0 and j < BOARD_COLS:
+            # If there is a solder in the space which doesn't belong to this player, return
+            if isinstance(space_list[i][j].get_unit(), Soldier): 
+                    if space_list[i][j].get_unit().get_player() != self.get_player():
+                        return valid_spaces 
+            # Otherwise proceed
+            valid_spaces = valid_spaces.union(self.find_move_spaces(i, j, range-1, space_list))
+        return valid_spaces
+    
 class Sorcerer(Unit):
     def __init__(self, p1 = True) -> None:
         hp=14
@@ -438,73 +473,6 @@ class Healer(Unit):
         if damage_healed > 0:
             attack_log.append(f"{unit_name} infuses {target_name} with mystical energy, healing {damage_healed} damage!\n")
         return attack_log
-
-class Archer(Unit):
-    def __init__(self, p1 = True) -> None:
-        hp=15
-        dam_val=6
-        dam_type=DamageType.PIERCE
-        arm_val=2
-        arm_type=ArmourType.PADDED
-        move=MoveSpeed.MED
-        move_type = MoveType.FOOT
-        if p1:
-            sprite = SpriteType.ARCHER1
-        else:
-            sprite = SpriteType.ARCHER2
-        name_list = Names.Commoner
-        title_list = Titles.Archer
-        ability_name = "Ranged Attack"
-        ability_range = 5
-        ability_value = 7
-        super().__init__(hp, dam_val, dam_type, arm_val, arm_type, move, move_type, sprite, name_list, title_list, ability_name, ability_range, ability_value)
-        self.__special_damage_type = DamageType.PIERCE
-
-    def special_ability(self, target, spaces):
-        unit_name = self.get_name()
-        target_name = target.get_name()
-        attack_log = []
-        first_strike_attack = ceil(self.get_ability_value() * FIRST_STRIKE_BOOST)
-        target_hp = target.get_curr_hp()
-        self.attack(target, first_strike_attack, self.__special_damage_type)
-        damage_dealt = target_hp - target.get_curr_hp()
-        attack_log.append(f"{unit_name} fires an arrow at {target_name}, dealing {damage_dealt} damage!\n")
-        if target.is_dead():
-            attack_log.append(f"{unit_name} has slain {target_name}!\n")
-            target.get_location().assign_unit(None)
-        return attack_log
-
-class Cavalry(Unit):
-    def __init__(self, p1 = True) -> None:
-        hp=20
-        dam_val=9
-        dam_type=DamageType.SLASH
-        arm_val=4
-        arm_type=ArmourType.PLATE
-        move=MoveSpeed.FAST
-        move_type = MoveType.HORSE
-        if p1:
-            sprite = SpriteType.PEASANT1
-        else:
-            sprite = SpriteType.PEASANT2
-        name_list = Names.Noble
-        title_list = Titles.Cavalry
-        ability_name = "Harrying Strike"
-        ability_range = 1
-        ability_value = None
-        super().__init__(hp, dam_val, dam_type, arm_val, arm_type, move, move_type, sprite, name_list, title_list, ability_name, ability_range, ability_value)
-    
-    def check_move_spaces(self, i: int, j: int, range: int, space_list: list) -> set:
-        valid_spaces = set()
-        if i >= 0 and i < BOARD_ROWS and j >= 0 and j < BOARD_COLS:
-            # If there is a solder in the space which doesn't belong to this player, return
-            if isinstance(space_list[i][j].get_unit(), Soldier): 
-                    if space_list[i][j].get_unit().get_player() != self.get_player():
-                        return valid_spaces 
-            # Otherwise proceed
-            valid_spaces = valid_spaces.union(self.find_move_spaces(i, j, range-1, space_list))
-        return valid_spaces
-    
 
 class Archmage(Unit):
     def __init__(self, p1 = True) -> None:
